+++ KetaminRudwolf/src/render/GUIRenderer.hpp
@@ -13,8 +13,8 @@ public:
 	static GUIRenderer* getInstance();
 	void newFrame();
 	void render();
-	void begin(const char* title, int x, int y, int width, int height);
-	void beginBorderedWindow(const char* title, int x, int y, int width, int height);
+	void begin(const char* title, float x, float y, float width, float height);
+	void beginBorderedWindow(const char* title, float x, float y, float width, float height);
 	void end();
 	void row(float height, int cols);
 	bool button(const char *title);
